abstracted components into sub directories & adjusted space weather notifications
