+++ src/components/space-weather-notifications/space-weather-notifications.js
@@ -1,6 +1,8 @@
 import { fetchSpaceWeatherAlerts } from "@/services/fetch-datasets";
 import { useEffect, useState } from "react";
-import SpaceWeatherModal from "./space-weather-modal";
+import Modal from "../modal";
+import ReactMarkdown from "react-markdown";
+import remarkGfm from "remark-gfm";
 
 export const SpaceWeatherNotifications = () => {
   const [space_weather_alerts, setSpaceWeatherAlerts] = useState([]);
@@ -38,9 +40,7 @@ export const SpaceWeatherNotifications = () => {
 
   return (
     <section>
-      <h2 className="text-2xl font-semibold mb-4">
-        Space Weather Notifications
-      </h2>
+      <h2 className="text-2xl font-semibold mb-4">Space Weather Notifications</h2>
       <div className="grid grid-cols-1 sm:grid-cols-2 gap-4">
         {(() => {
           // sort newest-first and slice for pagination
@@ -50,21 +50,19 @@ export const SpaceWeatherNotifications = () => {
           const startIdx = (currentPage - 1) * itemsPerPage;
           const pageAlerts = sorted.slice(startIdx, startIdx + itemsPerPage);
           return pageAlerts.map((alert) => (
-          <div
-            key={alert.message_id}
-            role="button"
-            tabIndex={0}
-            onClick={() => setSelectedAlert(alert)}
-            onKeyDown={(e) => e.key === "Enter" && setSelectedAlert(alert)}
-            className="bg-white dark:bg-gray-800 rounded-lg p-3 shadow cursor-pointer hover:shadow-md focus:shadow-md outline-none"
-          >
-            <div className="flex items-center justify-between">
-              <h3 className="font-medium text-sm truncate">{alert.message_type}</h3>
-              <div className="text-xs text-gray-600 dark:text-gray-300 ml-2">
-                {alert.message_issue_time}
+            <div
+              key={alert.message_id}
+              role="button"
+              tabIndex={0}
+              onClick={() => setSelectedAlert(alert)}
+              onKeyDown={(e) => e.key === "Enter" && setSelectedAlert(alert)}
+              className="bg-white dark:bg-gray-800 rounded-lg p-3 shadow cursor-pointer hover:shadow-md focus:shadow-md outline-none"
+            >
+              <div className="flex items-center justify-between">
+                <h3 className="font-medium text-sm truncate">{alert.message_type}</h3>
+                <div className="text-xs text-gray-600 dark:text-gray-300 ml-2">{alert.message_issue_time}</div>
               </div>
             </div>
-          </div>
           ));
         })()}
       </div>
@@ -86,9 +84,7 @@ export const SpaceWeatherNotifications = () => {
           </div>
 
           <button
-            onClick={() =>
-              setCurrentPage((p) => Math.min(Math.ceil(space_weather_alerts.length / itemsPerPage), p + 1))
-            }
+            onClick={() => setCurrentPage((p) => Math.min(Math.ceil(space_weather_alerts.length / itemsPerPage), p + 1))}
             disabled={currentPage >= Math.ceil(space_weather_alerts.length / itemsPerPage)}
             className="px-3 py-1 rounded bg-gray-100 dark:bg-gray-700 text-sm disabled:opacity-50"
             aria-label="Next page"
@@ -98,7 +94,41 @@ export const SpaceWeatherNotifications = () => {
         </div>
       )}
 
-  <SpaceWeatherModal selectedAlert={selectedAlert} onClose={() => setSelectedAlert(null)} />
+      {/* Modal */}
+      {selectedAlert && (
+        <Modal isOpen={!!selectedAlert} onClose={() => setSelectedAlert(null)} title={selectedAlert.message_type}>
+          <div className="text-sm text-gray-600 dark:text-gray-300">{selectedAlert.message_issue_time}</div>
+
+          <div className="mt-4">
+            <div className="prose dark:prose-invert max-w-full break-words">
+              <ReactMarkdown
+                remarkPlugins={[remarkGfm]}
+                components={{
+                  pre: ({ node, children, ...props }) => (
+                    <div className="overflow-auto">
+                      <pre className="whitespace-pre-wrap break-words" {...props}>
+                        {children}
+                      </pre>
+                    </div>
+                  ),
+                  code: ({ node, inline, className, children, ...props }) =>
+                    inline ? (
+                      <code className="whitespace-normal break-words" {...props}>
+                        {children}
+                      </code>
+                    ) : (
+                      <code className="block whitespace-pre-wrap break-words" {...props}>
+                        {children}
+                      </code>
+                    ),
+                }}
+              >
+                {selectedAlert.message_body}
+              </ReactMarkdown>
+            </div>
+          </div>
+        </Modal>
+      )}
     </section>
   );
 };
